[refact]Toc 기능 리팩토링
- 불필요하게 복잡한 IntersectionObserver 훅스 삭제.
- toc 기능 재구현.
